refactor article path
